requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
anyio=3.3.3=py38h578d9bd_0
argon2-cffi=20.1.0=py38h27cfd23_1
async_generator=1.10=py_0
attrs=21.2.0=pyhd8ed1ab_0
babel=2.9.1=pyh44b312d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bleach=4.1.0=pyhd8ed1ab_0
bottleneck=1.3.2=py38heb32a55_1
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py38h497a2fe_1001
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.10.11=h06a4308_0
certifi=2022.9.24=py38h06a4308_0
cffi=1.14.6=py38ha65f79e_0
chardet=4.0.0=py38h578d9bd_1
charset-normalizer=2.0.0=pyhd8ed1ab_0
cityflow=0.1=pypi_0
cryptography=35.0.0=py38ha5dfef3_0
cuda=11.6.2=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=11.8.86=0
cuda-libraries=11.6.2=0
cuda-libraries-dev=11.6.2=0
cuda-memcheck=11.8.86=0
cuda-nsight=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=11.8.87=0
cuda-runtime=11.6.2=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.6.2=0
cuda-tools=11.6.2=0
cuda-visual-tools=11.6.2=0
cudatoolkit=11.1.74=h6bb024c_0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
debugpy=1.4.1=py38h709712a_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
expat=2.4.1=h2531618_2
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h5ab3b9f_0
gds-tools=1.4.0.31=0
glib=2.69.1=h5202010_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.4=pyhd8ed1ab_0
importlib-metadata=4.8.1=py38h578d9bd_0
intel-openmp=2021.3.0=h06a4308_3350
ipykernel=6.4.1=py38he5a9106_0
ipython=7.28.0=py38he5a9106_0
ipython_genutils=0.2.0=py_1
jedi=0.18.0=py38h578d9bd_2
jinja2=3.0.2=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
json5=0.9.5=pyh9f0ad1d_0
jsonschema=4.1.0=pyhd8ed1ab_0
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=4.8.1=py38h578d9bd_0
jupyter_server=1.11.1=pyhd8ed1ab_0
jupyterlab=3.2.0=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.8.2=pyhd8ed1ab_0
kiwisolver=1.3.1=py38h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcufile=1.4.0.31=0
libcufile-dev=1.4.0.31=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h2531618_0
markupsafe=2.0.1=py38h497a2fe_0
matplotlib=3.4.2=py38h06a4308_0
matplotlib-base=3.4.2=py38hab158f2_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py38h497a2fe_1004
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
munkres=1.1.4=py_0
nbclassic=0.3.2=pyhd8ed1ab_0
nbclient=0.5.4=pyhd8ed1ab_0
nbconvert=6.2.0=py38h578d9bd_0
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=hff7bd54_1
notebook=6.4.4=pyha770c72_0
nsight-compute=2022.3.0.22=0
numexpr=2.7.3=py38h22e1b3c_1
numpy=1.20.3=py38hf144106_0
numpy-base=1.20.3=py38h74d4b33_0
olefile=0.46=py_0
openh264=2.1.0=hd408876_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1s=h7f8727e_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.3.3=py38h8c16a72_0
pandoc=2.14.2=h7f98852_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.2=pyhd8ed1ab_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.3.1=py38h2c7a002_0
pip=21.1.3=py38h06a4308_0
prometheus_client=0.11.0=pyhd8ed1ab_0
prompt-toolkit=3.0.20=pyha770c72_0
ptyprocess=0.7.0=pyhd3deb0d_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.10.0=pyhd8ed1ab_0
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.17.3=py38h497a2fe_2
pysocks=1.7.1=py38h578d9bd_3
python=3.8.10=h12debd9_8
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.10.2=py3.8_cuda11.1_cudnn8.0.5_0
pytorch-cuda=11.6=h867d48c_0
pytorch-mutex=1.0=cuda
pytz=2021.1=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
pyzmq=19.0.2=py38ha71036d_2
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.28.1=pyhd8ed1ab_1
requests-unixsocket=0.2.0=py_0
scipy=1.7.1=py38h292c36d_2
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=52.0.0=py38h06a4308_0
sip=4.19.13=py38he6710b0_0
six=1.16.0=pyhd3eb1b0_0
sniffio=1.2.0=py38h578d9bd_1
sqlite=3.36.0=hc218d9a_0
terminado=0.12.1=py38h578d9bd_0
testpath=0.5.0=pyhd8ed1ab_0
tk=8.6.10=hbc83047_0
torchaudio=0.10.2=py38_cu111
torchvision=0.11.3=py38_cu111
tornado=6.1=py38h27cfd23_0
traitlets=5.1.0=pyhd8ed1ab_0
typing_extensions=3.10.0.0=pyh06a4308_0
ujson=1.35=py38h7b6447c_0
urllib3=1.26.11=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=0.57.0=py38h578d9bd_4
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h9c3ff4c_0
zipp=3.6.0=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
